Update PyTorch setup for tf32 and change default WMT workload compute dtype to FLOAT32

+++ algoperf/workloads/wmt/workload.py
@@ -22,7 +22,7 @@ class BaseWmtWorkload(spec.Workload):
   """A WMT workload."""
 
   _vocab_size: int = 32000
-  _compute_dtype: spec.DTYPE = spec.DTYPE.BFLOAT16
+  _compute_dtype: spec.DTYPE = spec.DTYPE.FLOAT32
   _param_dtype: spec.DTYPE = spec.DTYPE.FLOAT32
 
   def __init__(self) -> None:
